portfolio_intelligence/backend/requirements.txt
# FastAPI + Server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0

# HTTP Client
httpx>=0.27.0

# Config & Validation
pydantic-settings>=2.1.0

# Auth
python-jose[cryptography]>=3.3.0
cryptography>=42.0.0

# Task Queue & Cache
celery[redis]>=5.3.0
redis>=5.0.0
APScheduler>=3.10.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx  # also used as test client

# Utils
python-multipart>=0.0.6
